// CidtcaFrontend/src/interfaces/formsToExcel.ts
export interface FormsToExcelInterface {
  pregunta_1: string;
  pregunta_2: string;
  pregunta_2_otro: string;
  pregunta_3: string;
  pregunta_4: string;
  pregunta_5_numero: string;
  pregunta_5_correo: string;
  pregunta_6: string;
  pregunta_7: string;
  pregunta_8: string;
  pregunta_9: string;
  pregunta_9_otro: string;
  pregunta_10: string;
  pregunta_10_otro: string;
  pregunta_11: string;
  pregunta_11_otro: string;
  pregunta_12: string;
  pregunta_12_otro: string;
  pregunta_13: string;
  pregunta_14: string;
  pregunta_15: string;
  pregunta_16: string;
  pregunta_16_cuantos: string;
  pregunta_17: string;
  pregunta_17_cuantos: string;
  pregunta_18: string;
  pregunta_18_cuantos: string;
  pregunta_19: string;
  pregunta_20: string;
  pregunta_21: string;
  pregunta_23: string;
  pregunta_23_gps: string;
  pregunta_24: string;
  pregunta_25: string;
  pregunta_26: string;
  pregunta_27: string;
  pregunta_28_mensual: string;
  pregunta_29_yogurt_diaria: string;
  pregunta_29_yogurt_semanal: string;
  pregunta_29_yogurt_mensual: string;
  pregunta_29_kumis_diaria: string;
  pregunta_29_kumis_semanal: string;
  pregunta_29_kumis_mensual: string;
  pregunta_29_cuajada_diaria: string;
  pregunta_29_cuajada_semanal: string;
  pregunta_29_cuajada_mensual: string;
  pregunta_29_queso_doble_crema_diaria: string;
  pregunta_29_queso_doble_crema_semanal: string;
  pregunta_29_queso_doble_crema_mensual: string;
  pregunta_29_queso_campesino_diaria: string;
  pregunta_29_queso_campesino_semanal: string;
  pregunta_29_queso_campesino_mensual: string;
  pregunta_29_leche_diaria: string;
  pregunta_29_leche_semanal: string;
  pregunta_29_leche_mensual: string;
  pregunta_29_quesos_madurados_diaria: string;
  pregunta_29_quesos_madurados_semanal: string;
  pregunta_29_quesos_madurados_mensual: string;
  pregunta_29_requeson_diaria: string;
  pregunta_29_requeson_semanal: string;
  pregunta_29_requeson_mensual: string;
  pregunta_29_arequipe_diaria: string;
  pregunta_29_arequipe_semanal: string;
  pregunta_29_arequipe_mensual: string;
  pregunta_29_otros: string;
  pregunta_30: string;
  pregunta_31: string;
  pregunta_32: string;
  pregunta_33: string;
  pregunta_34: string;
  pregunta_35: string;
  pregunta_36: string;
  pregunta_37: string;
  pregunta_37_otro: string;
  pregunta_38: string;
  pregunta_38_otro: string;
  pregunta_39: string;
  pregunta_39_otro: string;
  pregunta_40: string;
  pregunta_40_cuales: string;
  pregunta_41: string;
  pregunta_42: string;
  pregunta_43: string;
  pregunta_43_cuales: string;
  pregunta_44: string;
  pregunta_44_como: string;
  pregunta_45: string;
  pregunta_45_como: string;
  pregunta_46: string;
  pregunta_47: string;
  pregunta_48: string;
  pregunta_48_cual: string;
  pregunta_49: string;
  pregunta_50: string;
  pregunta_51: string;
  pregunta_52: string;
  pregunta_53: string;
  pregunta_54: string;
  pregunta_55: string;
  pregunta_56: string;
  pregunta_56_cuales: string;
  pregunta_57: string;
  pregunta_57_destino_transformadora_local: string;
  pregunta_57_destino_transformadora_nacional: string;
  pregunta_57_destino_venta_local: string;
  pregunta_57_destino_venta_nacional: string;
  pregunta_57_otros: string;
  pregunta_58: string;
  pregunta_59: string;
  pregunta_60: string;
  pregunta_60_otros: string;
  pregunta_61: string;
  pregunta_62: string;
  pregunta_63: string;
  pregunta_64_energia: string;
  pregunta_64_gas: string;
  pregunta_64_otro: string;
  pregunta_64_otro_valor: string;
  pregunta_65: string;
  pregunta_65_porque: string;
  pregunta_66_yogurt_presentacion: string;
  pregunta_66_yogurt_valor: string;
  pregunta_66_kumis_presentacion: string;
  pregunta_66_kumis_valor: string;
  pregunta_66_cuajada_presentacion: string;
  pregunta_66_cuajada_valor: string;
  pregunta_66_queso_doble_crema_presentacion: string;
  pregunta_66_queso_doble_crema_valor: string;
  pregunta_66_queso_campesino_presentacion: string;
  pregunta_66_queso_campesino_valor: string;
  pregunta_66_leche_presentacion: string;
  pregunta_66_leche_valor: string;
  pregunta_66_quesos_madurados_presentacion: string;
  pregunta_66_quesos_madurados_valor: string;
  pregunta_66_requeson_presentacion: string;
  pregunta_66_requeson_valor: string;
  pregunta_66_arequipe_presentacion: string;
  pregunta_66_arequipe_valor: string;
  pregunta_66_otros: string;
  observaciones: string;
  fecha: string;
  nombre_encuestado: string;
  cedula_encuestado: string;
  firma_encuestado: string;
  nombre_encuestador: string;
  cedula_encuestador: string;
  firma_encuestador: string;
  foto_encuestado: string;
}